requirements.txt
streamlit==1.30.0
requests==2.31.0
tmdbv3api==1.9.0
trakt==3.4.0
requests-oauthlib==1.3.1
rapidfuzz==3.6.1
pydantic==2.5.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
loguru==0.7.2
pandas==2.1.4
pyyaml==6.0.1
pytest==7.4.3
